[lisp] Moving //lisp/core and //lisp/test to third_party.
This is just moving the source.
Only modifications necessary for the move have been performed.
Some fix to documentation was done to reflect new paths.
It is in preparation for staging the source on an internal git repository
and then for open-sourcing those libraries.

Further modifications and cleanup needs to be performed.

PiperOrigin-RevId: 310628989